stats242-executable-textbook-v0.6/requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
statsmodels>=0.14
matplotlib>=3.7
pyyaml>=6.0
markdown>=3.5
pytest>=7.4
